rr_graph: Split testarch_graph out of graph.py
